♻️  zv: Split out zvariant tests into separate files
under `tests` directory.
