+++ zvariant/src/lib.rs
@@ -108,1845 +108,3 @@ pub mod export {
 
 // Re-export all of the `endi` API for ease of use.
 pub use endi::*;
-
-#[cfg(test)]
-#[allow(clippy::disallowed_names)]
-mod tests {
-    use std::collections::{BTreeMap, HashMap};
-
-    #[cfg(feature = "arrayvec")]
-    use arrayvec::{ArrayString, ArrayVec};
-    use serde_json::json;
-    #[cfg(feature = "arrayvec")]
-    use std::str::FromStr;
-    use zvariant_utils::signature::Signature;
-
-    use serde::{Deserialize, Serialize};
-
-    use crate::{
-        as_value::{self, optional},
-        to_bytes, to_bytes_for_signature, MaxDepthExceeded,
-    };
-
-    #[cfg(unix)]
-    use crate::Fd;
-    use crate::{
-        serialized::{Context, Format},
-        Array, Basic, Dict, Error, ObjectPath, OwnedValue, Result, Str, Structure, Type, Value, BE,
-        LE, NATIVE_ENDIAN,
-    };
-
-    // Test through both generic and specific API (wrt byte order)
-    macro_rules! basic_type_test {
-        ($endian:expr, $format:ident, $test_value:expr, $expected_len:expr, $expected_ty:ty, $align:literal) => {{
-            // Lie that we're starting at byte 1 in the overall message to test padding
-            let ctxt = Context::new(Format::$format, $endian, 1);
-            let encoded = to_bytes(ctxt, &$test_value).unwrap();
-            let padding = crate::padding_for_n_bytes(1, $align);
-            assert_eq!(
-                encoded.len(),
-                $expected_len + padding,
-                "invalid encoding using `to_bytes`"
-            );
-            let (decoded, parsed): ($expected_ty, _) = encoded.deserialize().unwrap();
-            assert!(decoded == $test_value, "invalid decoding");
-            assert!(parsed == encoded.len(), "invalid parsing");
-
-            // Now encode w/o padding
-            let ctxt = Context::new(Format::$format, $endian, 0);
-            let encoded = to_bytes(ctxt, &$test_value).unwrap();
-            assert_eq!(
-                encoded.len(),
-                $expected_len,
-                "invalid encoding using `to_bytes`"
-            );
-
-            encoded
-        }};
-        ($endian:expr, $format:ident, $test_value:expr, $expected_len:expr, $expected_ty:ty, $align:literal, $kind:ident, $expected_value_len:expr) => {{
-            let encoded = basic_type_test!(
-                $endian,
-                $format,
-                $test_value,
-                $expected_len,
-                $expected_ty,
-                $align
-            );
-
-            // As Value
-            let v: Value<'_> = $test_value.into();
-            assert_eq!(v.value_signature(), <$expected_ty>::SIGNATURE_STR);
-            assert_eq!(v, Value::$kind($test_value));
-            value_test!(LE, $format, v, $expected_value_len);
-
-            let v: $expected_ty = v.try_into().unwrap();
-            assert_eq!(v, $test_value);
-
-            encoded
-        }};
-    }
-
-    macro_rules! value_test {
-        ($endian:expr, $format:ident, $test_value:expr, $expected_len:expr) => {{
-            let ctxt = Context::new(Format::$format, $endian, 0);
-            let encoded = to_bytes(ctxt, &$test_value).unwrap();
-            assert_eq!(
-                encoded.len(),
-                $expected_len,
-                "invalid encoding using `to_bytes`"
-            );
-            let (decoded, parsed): (Value<'_>, _) = encoded.deserialize().unwrap();
-            assert!(decoded == $test_value, "invalid decoding");
-            assert!(parsed == encoded.len(), "invalid parsing");
-
-            encoded
-        }};
-    }
-
-    fn f64_type_test(
-        format: Format,
-        value: f64,
-        expected_len: usize,
-        expected_value_len: usize,
-    ) -> crate::serialized::Data<'static, 'static> {
-        // Lie that we're starting at byte 1 in the overall message to test padding
-        let ctxt = Context::new(format, NATIVE_ENDIAN, 1);
-        let encoded = to_bytes(ctxt, &value).unwrap();
-        let padding = crate::padding_for_n_bytes(1, 8);
-        assert_eq!(
-            encoded.len(),
-            expected_len + padding,
-            "invalid encoding using `to_bytes`"
-        );
-
-        let decoded: f64 = encoded.deserialize().unwrap().0;
-        assert!(
-            (decoded - value).abs() < f64::EPSILON,
-            "invalid decoding using `from_slice`"
-        );
-
-        // Now encode w/o padding
-        let ctxt = Context::new(format, NATIVE_ENDIAN, 0);
-        let encoded = to_bytes(ctxt, &value).unwrap();
-        assert_eq!(
-            encoded.len(),
-            expected_len,
-            "invalid encoding using `to_bytes`"
-        );
-
-        f64_type_test_as_value(format, value, expected_value_len);
-        encoded
-    }
-
-    fn f64_type_test_as_value(format: Format, value: f64, expected_value_len: usize) {
-        let v: Value<'_> = value.into();
-        assert_eq!(v.value_signature(), f64::SIGNATURE_STR);
-        assert_eq!(v, Value::F64(value));
-        f64_value_test(format, v.try_clone().unwrap(), expected_value_len);
-        let v: f64 = v.try_into().unwrap();
-        assert!((v - value).abs() < f64::EPSILON);
-    }
-
-    fn f64_value_test(format: Format, v: Value<'_>, expected_value_len: usize) {
-        let ctxt = Context::new(format, LE, 0);
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(
-            encoded.len(),
-            expected_value_len,
-            "invalid encoding using `to_bytes`"
-        );
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        assert!(decoded == v, "invalid decoding using `from_slice`");
-    }
-
-    // All fixed size types have the same encoding in DBus and GVariant formats.
-    //
-    // NB: Value (i-e VARIANT type) isn't a fixed size type.
-
-    #[test]
-    fn u8_value() {
-        let encoded = basic_type_test!(LE, DBus, 77_u8, 1, u8, 1, U8, 4);
-        assert_eq!(encoded.len(), 1);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, 77_u8, 1, u8, 1, U8, 3);
-    }
-
-    #[test]
-    fn i8_value() {
-        basic_type_test!(LE, DBus, 77_i8, 2, i8, 2);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, 77_i8, 2, i8, 2);
-    }
-
-    #[cfg(unix)]
-    macro_rules! fd_value_test {
-        ($endian:expr, $format:ident, $test_value:expr, $expected_len:expr, $align:literal, $expected_value_len:expr) => {{
-            use std::os::fd::AsFd;
-
-            // Lie that we're starting at byte 1 in the overall message to test padding
-            let ctxt = Context::new(Format::$format, $endian, 1);
-            let encoded = to_bytes(ctxt, &$test_value).unwrap();
-            let padding = crate::padding_for_n_bytes(1, $align);
-            assert_eq!(
-                encoded.len(),
-                $expected_len + padding,
-                "invalid encoding using `to_bytes`"
-            );
-            #[cfg(unix)]
-            let (_, parsed): (Fd<'_>, _) = encoded.deserialize().unwrap();
-            assert!(
-                parsed == encoded.len(),
-                "invalid parsing using `from_slice`"
-            );
-
-            // Now encode w/o padding
-            let ctxt = Context::new(Format::$format, $endian, 0);
-            let encoded = to_bytes(ctxt, &$test_value).unwrap();
-            assert_eq!(
-                encoded.len(),
-                $expected_len,
-                "invalid encoding using `to_bytes`"
-            );
-
-            // As Value
-            let v: Value<'_> = $test_value.into();
-            assert_eq!(v.value_signature(), Fd::SIGNATURE_STR);
-            assert_eq!(v, Value::Fd($test_value));
-            let encoded = to_bytes(ctxt, &v).unwrap();
-            assert_eq!(encoded.fds().len(), 1, "invalid encoding using `to_bytes`");
-            assert_eq!(
-                encoded.len(),
-                $expected_value_len,
-                "invalid encoding using `to_bytes`"
-            );
-            let (decoded, parsed): (Value<'_>, _) = encoded.deserialize().unwrap();
-            assert_eq!(
-                decoded,
-                Fd::from(encoded.fds()[0].as_fd()).into(),
-                "invalid decoding using `from_slice`"
-            );
-            assert_eq!(parsed, encoded.len(), "invalid parsing using `from_slice`");
-
-            let v: Fd<'_> = v.try_into().unwrap();
-            assert_eq!(v, $test_value);
-        }};
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn fd_value() {
-        use std::os::fd::AsFd;
-
-        let stdout = std::io::stdout();
-        let fd = stdout.as_fd();
-        fd_value_test!(LE, DBus, Fd::from(fd), 4, 4, 8);
-        #[cfg(feature = "gvariant")]
-        fd_value_test!(LE, GVariant, Fd::from(fd), 4, 4, 6);
-    }
-
-    #[test]
-    fn u16_value() {
-        let encoded = basic_type_test!(BE, DBus, 0xABBA_u16, 2, u16, 2, U16, 6);
-        assert_eq!(encoded.len(), 2);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(BE, GVariant, 0xABBA_u16, 2, u16, 2, U16, 4);
-    }
-
-    #[test]
-    fn i16_value() {
-        let encoded = basic_type_test!(BE, DBus, -0xAB0_i16, 2, i16, 2, I16, 6);
-        assert_eq!(LE.read_i16(&encoded), 0x50F5_i16);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(BE, GVariant, -0xAB0_i16, 2, i16, 2, I16, 4);
-    }
-
-    #[test]
-    fn u32_value() {
-        let encoded = basic_type_test!(BE, DBus, 0xABBA_ABBA_u32, 4, u32, 4, U32, 8);
-        assert_eq!(encoded.len(), 4);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(BE, GVariant, 0xABBA_ABBA_u32, 4, u32, 4, U32, 6);
-    }
-
-    #[test]
-    fn i32_value() {
-        let encoded = basic_type_test!(BE, DBus, -0xABBA_AB0_i32, 4, i32, 4, I32, 8);
-        assert_eq!(LE.read_i32(&encoded), 0x5055_44F5_i32);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(BE, GVariant, -0xABBA_AB0_i32, 4, i32, 4, I32, 6);
-    }
-
-    // u64 is covered by `value_value` test below
-
-    #[test]
-    fn i64_value() {
-        let encoded = basic_type_test!(BE, DBus, -0xABBA_ABBA_ABBA_AB0_i64, 8, i64, 8, I64, 16);
-        assert_eq!(LE.read_i64(&encoded), 0x5055_4455_4455_44F5_i64);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(BE, GVariant, -0xABBA_ABBA_ABBA_AB0_i64, 8, i64, 8, I64, 10);
-    }
-
-    #[test]
-    fn f64_value() {
-        let encoded = f64_type_test(Format::DBus, 99999.99999_f64, 8, 16);
-        assert!((NATIVE_ENDIAN.read_f64(&encoded) - 99999.99999_f64).abs() < f64::EPSILON);
-        #[cfg(feature = "gvariant")]
-        f64_type_test(Format::GVariant, 99999.99999_f64, 8, 10);
-    }
-
-    #[test]
-    fn str_value() {
-        let string = String::from("hello world");
-        basic_type_test!(LE, DBus, string, 16, String, 4);
-        basic_type_test!(LE, DBus, string, 16, &str, 4);
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, string, 12, String, 1);
-
-        let string = "hello world";
-        basic_type_test!(LE, DBus, string, 16, &str, 4);
-        basic_type_test!(LE, DBus, string, 16, String, 4);
-
-        // As Value
-        let v: Value<'_> = string.into();
-        assert_eq!(v.value_signature(), "s");
-        assert_eq!(v, Value::new("hello world"));
-        value_test!(LE, DBus, v, 20);
-        #[cfg(feature = "gvariant")]
-        value_test!(LE, GVariant, v, 14);
-
-        let v: String = v.try_into().unwrap();
-        assert_eq!(v, "hello world");
-
-        // Characters are treated as strings
-        basic_type_test!(LE, DBus, 'c', 6, char, 4);
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, 'c', 2, char, 1);
-
-        // As Value
-        let v: Value<'_> = "c".into();
-        assert_eq!(v.value_signature(), "s");
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 10);
-        let (v, _) = encoded.deserialize::<Value<'_>>().unwrap();
-        assert_eq!(v, Value::new("c"));
-    }
-
-    #[cfg(feature = "arrayvec")]
-    #[test]
-    fn array_string_value() {
-        let s = ArrayString::<32>::from_str("hello world!").unwrap();
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 17);
-        let decoded: ArrayString<32> = encoded.deserialize().unwrap().0;
-        assert_eq!(&decoded, "hello world!");
-    }
-
-    #[cfg(feature = "heapless")]
-    #[test]
-    fn heapless_string_value() {
-        use heapless::String;
-
-        let s = String::<32>::try_from("hello world!").unwrap();
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 17);
-        let decoded: String<32> = encoded.deserialize().unwrap().0;
-        assert_eq!(&decoded, "hello world!");
-    }
-
-    #[test]
-    fn signature() {
-        use crate::Signature;
-        use std::str::FromStr;
-
-        let sig = Signature::from_str("yys").unwrap();
-        // Structure will always add () around the signature if it's a struct.
-        basic_type_test!(LE, DBus, sig, 7, Signature, 1);
-
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, sig, 6, Signature, 1);
-
-        // As Value
-        let v: Value<'_> = sig.into();
-        assert_eq!(v.value_signature(), "g");
-        let encoded = value_test!(LE, DBus, v, 10);
-        let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-        assert_eq!(v, Value::Signature(Signature::try_from("yys").unwrap()));
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let encoded = value_test!(LE, GVariant, v, 8);
-            let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-            assert_eq!(v, Value::Signature(Signature::try_from("yys").unwrap()));
-        }
-    }
-
-    #[test]
-    fn object_path_value() {
-        let o = ObjectPath::try_from("/hello/world").unwrap();
-        basic_type_test!(LE, DBus, o, 17, ObjectPath<'_>, 4);
-
-        #[cfg(feature = "gvariant")]
-        basic_type_test!(LE, GVariant, o, 13, ObjectPath<'_>, 1);
-
-        // As Value
-        let v: Value<'_> = o.into();
-        assert_eq!(v.value_signature(), "o");
-        let encoded = value_test!(LE, DBus, v, 21);
-        let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-        assert_eq!(
-            v,
-            Value::ObjectPath(ObjectPath::try_from("/hello/world").unwrap())
-        );
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let encoded = value_test!(LE, GVariant, v, 15);
-            let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-            assert_eq!(
-                v,
-                Value::ObjectPath(ObjectPath::try_from("/hello/world").unwrap())
-            );
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn unit_fds() {
-        let ctxt = Context::new_dbus(BE, 0);
-        let encoded = to_bytes(ctxt, &()).unwrap();
-        assert_eq!(encoded.len(), 0, "invalid encoding using `to_bytes`");
-        let _: () = encoded
-            .deserialize()
-            .expect("invalid decoding using `from_slice`")
-            .0;
-    }
-
-    #[test]
-    fn unit() {
-        let ctxt = Context::new_dbus(BE, 0);
-        let encoded = to_bytes(ctxt, &()).unwrap();
-        assert_eq!(encoded.len(), 0, "invalid encoding using `to_bytes`");
-        let _: () = encoded
-            .deserialize()
-            .expect("invalid decoding using `from_slice`")
-            .0;
-    }
-
-    #[test]
-    fn array_value() {
-        // Let's use D-Bus/GVariant terms
-
-        //
-        // Array of u8
-        //
-        // First a normal Rust array that is actually serialized as a struct (thank you Serde!)
-        assert_eq!(<[u8; 2]>::SIGNATURE, "(yy)");
-        let ay = [77u8, 88];
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &ay).unwrap();
-        assert_eq!(encoded.len(), 2);
-        let decoded: [u8; 2] = encoded.deserialize().unwrap().0;
-        assert_eq!(&decoded, &[77u8, 88]);
-
-        // Then rest of the tests just use ArrayVec, heapless::Vec or Vec
-        #[cfg(feature = "arrayvec")]
-        let ay = ArrayVec::from([77u8, 88]);
-        #[cfg(all(not(feature = "arrayvec"), feature = "heapless"))]
-        let ay = heapless::Vec::<_, 2>::from_slice(&[77u8, 88]).unwrap();
-        #[cfg(all(not(feature = "arrayvec"), not(feature = "heapless")))]
-        let ay = vec![77u8, 88];
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &ay).unwrap();
-        assert_eq!(encoded.len(), 6);
-
-        #[cfg(feature = "arrayvec")]
-        let decoded: ArrayVec<u8, 2> = encoded.deserialize().unwrap().0;
-        #[cfg(all(not(feature = "arrayvec"), feature = "heapless"))]
-        let decoded: heapless::Vec<u8, 2> = encoded.deserialize().unwrap().0;
-        #[cfg(all(not(feature = "arrayvec"), not(feature = "heapless")))]
-        let decoded: Vec<u8> = encoded.deserialize().unwrap().0;
-        assert_eq!(&decoded.as_slice(), &[77u8, 88]);
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let gv_encoded = to_bytes(ctxt, &ay).unwrap();
-            assert_eq!(gv_encoded.len(), 2);
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // As Value
-        let v: Value<'_> = ay[..].into();
-        assert_eq!(v.value_signature(), "ay");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 10);
-        let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-        if let Value::Array(array) = v {
-            assert_eq!(*array.element_signature(), "y");
-            assert_eq!(array.len(), 2);
-            assert_eq!(array.get(0).unwrap(), Some(77u8));
-            assert_eq!(array.get(1).unwrap(), Some(88u8));
-        } else {
-            panic!();
-        }
-
-        // Now try as Vec
-        let vec = ay.to_vec();
-        let encoded = to_bytes(ctxt, &vec).unwrap();
-        assert_eq!(encoded.len(), 6);
-
-        // Vec as Value
-        let v: Value<'_> = Array::from(&vec).into();
-        assert_eq!(v.value_signature(), "ay");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 10);
-
-        // Empty array
-        let at: Vec<u64> = vec![];
-        let encoded = to_bytes(ctxt, &at).unwrap();
-        assert_eq!(encoded.len(), 8);
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let gv_encoded = to_bytes(ctxt, &at).unwrap();
-            assert_eq!(gv_encoded.len(), 0);
-            let at = encoded.deserialize::<Vec<u64>>().unwrap().0;
-            assert_eq!(at.len(), 0);
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // As Value
-        let v: Value<'_> = at[..].into();
-        assert_eq!(v.value_signature(), "at");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 8);
-        let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-        if let Value::Array(array) = v {
-            assert_eq!(*array.element_signature(), "t");
-            assert_eq!(array.len(), 0);
-        } else {
-            panic!();
-        }
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let v: Value<'_> = at[..].into();
-            let gv_encoded = to_bytes(ctxt, &v).unwrap();
-            assert_eq!(gv_encoded.len(), 3);
-            let v = gv_encoded.deserialize::<Value<'_>>().unwrap().0;
-            if let Value::Array(array) = v {
-                assert_eq!(*array.element_signature(), "t");
-                assert_eq!(array.len(), 0);
-            } else {
-                panic!();
-            }
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        //
-        // Array of strings
-        //
-        // Can't use 'as' as it's a keyword
-        let as_ = vec!["Hello", "World", "Now", "Bye!"];
-        let encoded = to_bytes(ctxt, &as_).unwrap();
-        assert_eq!(encoded.len(), 45);
-        let decoded = encoded.deserialize::<Vec<&str>>().unwrap().0;
-        assert_eq!(decoded.len(), 4);
-        assert_eq!(decoded[0], "Hello");
-        assert_eq!(decoded[1], "World");
-
-        let decoded = encoded.deserialize::<Vec<String>>().unwrap().0;
-        assert_eq!(decoded.as_slice(), as_.as_slice());
-
-        // Decode just the second string
-        let slice = encoded.slice(14..);
-        let decoded: &str = slice.deserialize().unwrap().0;
-        assert_eq!(decoded, "World");
-
-        // As Value
-        let v: Value<'_> = as_[..].into();
-        assert_eq!(v.value_signature(), "as");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 49);
-        let v = encoded.deserialize().unwrap().0;
-        if let Value::Array(array) = v {
-            assert_eq!(*array.element_signature(), "s");
-            assert_eq!(array.len(), 4);
-            assert_eq!(array[0], Value::new("Hello"));
-            assert_eq!(array[1], Value::new("World"));
-        } else {
-            panic!();
-        }
-
-        let v: Value<'_> = as_[..].into();
-        let a: Array<'_> = v.try_into().unwrap();
-        let _ve: Vec<String> = a.try_into().unwrap();
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let v: Value<'_> = as_[..].into();
-            let gv_encoded = to_bytes(ctxt, &v).unwrap();
-            assert_eq!(gv_encoded.len(), 28);
-        }
-
-        // Array of Struct, which in turn containin an Array (We gotta go deeper!)
-        // Signature: "a(yu(xbxas)s)");
-        let ar = vec![(
-            // top-most simple fields
-            u8::MAX,
-            u32::MAX,
-            (
-                // 2nd level simple fields
-                i64::MAX,
-                true,
-                i64::MAX,
-                // 2nd level array field
-                &["Hello", "World"][..],
-            ),
-            // one more top-most simple field
-            "hello",
-        )];
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &ar).unwrap();
-        assert_eq!(encoded.len(), 78);
-        #[allow(clippy::type_complexity)]
-        let decoded: Vec<(u8, u32, (i64, bool, i64, Vec<&str>), &str)> =
-            encoded.deserialize().unwrap().0;
-        assert_eq!(decoded.len(), 1);
-        let r = &decoded[0];
-        assert_eq!(r.0, u8::MAX);
-        assert_eq!(r.1, u32::MAX);
-        let inner_r = &r.2;
-        assert_eq!(inner_r.0, i64::MAX);
-        assert!(inner_r.1);
-        assert_eq!(inner_r.2, i64::MAX);
-        let as_ = &inner_r.3;
-        assert_eq!(as_.len(), 2);
-        assert_eq!(as_[0], "Hello");
-        assert_eq!(as_[1], "World");
-        assert_eq!(r.3, "hello");
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let gv_encoded = to_bytes(ctxt, &ar).unwrap();
-            assert_eq!(gv_encoded.len(), 54);
-            let decoded: Vec<(u8, u32, (i64, bool, i64, Vec<&str>), &str)> =
-                gv_encoded.deserialize().unwrap().0;
-            assert_eq!(decoded.len(), 1);
-            let r = &decoded[0];
-            assert_eq!(r.0, u8::MAX);
-            assert_eq!(r.1, u32::MAX);
-            let inner_r = &r.2;
-            assert_eq!(inner_r.0, i64::MAX);
-            assert!(inner_r.1);
-            assert_eq!(inner_r.2, i64::MAX);
-            let as_ = &inner_r.3;
-            assert_eq!(as_.len(), 2);
-            assert_eq!(as_[0], "Hello");
-            assert_eq!(as_[1], "World");
-            assert_eq!(r.3, "hello");
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // As Value
-        let v: Value<'_> = ar[..].into();
-        assert_eq!(v.value_signature(), "a(yu(xbxas)s)");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 94);
-        let v = encoded.deserialize::<Value<'_>>().unwrap().0;
-        if let Value::Array(array) = v.try_clone().unwrap() {
-            assert_eq!(*array.element_signature(), "(yu(xbxas)s)");
-            assert_eq!(array.len(), 1);
-            let r = &array[0];
-            if let Value::Structure(r) = r {
-                let fields = r.fields();
-                assert_eq!(fields[0], Value::U8(u8::MAX));
-                assert_eq!(fields[1], Value::U32(u32::MAX));
-                if let Value::Structure(r) = &fields[2] {
-                    let fields = r.fields();
-                    assert_eq!(fields[0], Value::I64(i64::MAX));
-                    assert_eq!(fields[1], Value::Bool(true));
-                    assert_eq!(fields[2], Value::I64(i64::MAX));
-                    if let Value::Array(as_) = &fields[3] {
-                        assert_eq!(as_.len(), 2);
-                        assert_eq!(as_[0], Value::new("Hello"));
-                        assert_eq!(as_[1], Value::new("World"));
-                    } else {
-                        panic!();
-                    }
-                } else {
-                    panic!();
-                }
-                assert_eq!(fields[3], Value::new("hello"));
-            } else {
-                panic!();
-            }
-        } else {
-            panic!();
-        }
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            use rand::{distr::Alphanumeric, rng, Rng};
-
-            let ctxt = Context::new_gvariant(LE, 0);
-            let gv_encoded = to_bytes(ctxt, &v).unwrap();
-            assert_eq!(gv_encoded.len(), 68);
-            let v: Value<'_> = gv_encoded.deserialize().unwrap().0;
-            if let Value::Array(array) = v {
-                assert_eq!(*array.element_signature(), "(yu(xbxas)s)");
-                assert_eq!(array.len(), 1);
-                let r = &array.get(0).unwrap().unwrap();
-                if let Value::Structure(r) = r {
-                    let fields = r.fields();
-                    assert_eq!(fields[0], Value::U8(u8::MAX));
-                    assert_eq!(fields[1], Value::U32(u32::MAX));
-                    if let Value::Structure(r) = &fields[2] {
-                        let fields = r.fields();
-                        assert_eq!(fields[0], Value::I64(i64::MAX));
-                        assert_eq!(fields[1], Value::Bool(true));
-                        assert_eq!(fields[2], Value::I64(i64::MAX));
-                        if let Value::Array(as_) = &fields[3] {
-                            assert_eq!(as_.len(), 2);
-                            assert_eq!(as_.get(0).unwrap(), Some("Hello"));
-                            assert_eq!(as_.get(1).unwrap(), Some("World"));
-                        } else {
-                            panic!();
-                        }
-                    } else {
-                        panic!();
-                    }
-                    assert_eq!(fields[3], Value::new("hello"));
-                } else {
-                    panic!();
-                }
-            } else {
-                panic!();
-            }
-
-            let mut rng = rng();
-            // Let's test GVariant ser/de of a 254 byte array with variable-width elements as to
-            // ensure no problems with non-normal BS of GVariant.
-            let as_ = vec![
-                (&mut rng)
-                    .sample_iter(Alphanumeric)
-                    .map(char::from)
-                    .take(126)
-                    .collect::<String>(),
-                (&mut rng)
-                    .sample_iter(Alphanumeric)
-                    .map(char::from)
-                    .take(126)
-                    .collect::<String>(),
-            ];
-            let gv_encoded = to_bytes(ctxt, &as_).unwrap();
-            // 252 chars + 2 null terminator bytes doesn't leave room for 2 framing offset bytes so
-            // a 2-byte offset is chosen by the serializer.
-            assert_eq!(gv_encoded.len(), 258);
-
-            // Check if our own deserializer does the right thing
-            let as2: Vec<String> = gv_encoded.deserialize().unwrap().0;
-            assert_eq!(as2, as_);
-
-            // Test conversion of Array of Value to Vec<Value>
-            let v = Value::new(vec![Value::new(43), Value::new("bonjour")]);
-            let av = <Array<'_>>::try_from(v).unwrap();
-            let av = <Vec<Value<'_>>>::try_from(av).unwrap();
-            assert_eq!(av[0], Value::new(43));
-            assert_eq!(av[1], Value::new("bonjour"));
-
-            let vec = vec![1, 2];
-            let val = Value::new(&vec);
-            assert_eq!(TryInto::<Vec<i32>>::try_into(val).unwrap(), vec);
-        }
-
-        // Empty array should be treated as a unit type, which is encoded as a u8.
-        assert_eq!(<[u64; 0]>::SIGNATURE, &Signature::U8);
-        let array: [u64; 0] = [];
-        let encoded = to_bytes(ctxt, &array).unwrap();
-        assert_eq!(encoded.len(), 1);
-        assert_eq!(encoded[0], 0);
-        let _decoded: [u64; 0] = encoded.deserialize().unwrap().0;
-    }
-
-    #[test]
-    fn struct_byte_array() {
-        let ctxt = Context::new_dbus(LE, 0);
-        let value: (Vec<u8>, HashMap<String, Value<'_>>) = (Vec::new(), HashMap::new());
-        let value = zvariant::to_bytes(ctxt, &value).unwrap();
-        #[cfg(feature = "serde_bytes")]
-        let (bytes, map): (&serde_bytes::Bytes, HashMap<&str, Value<'_>>) = value
-            .deserialize()
-            .expect("Could not deserialize serde_bytes::Bytes in struct.")
-            .0;
-        #[cfg(not(feature = "serde_bytes"))]
-        let (bytes, map): (&[u8], HashMap<&str, Value<'_>>) = value
-            .deserialize()
-            .expect("Could not deserialize u8 slice in struct")
-            .0;
-
-        assert!(bytes.is_empty());
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn struct_value() {
-        // Struct->Value
-        let s: Value<'_> = ("a", "b", (1, 2)).into();
-
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(dbg!(encoded.len()), 40);
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        let s = <Structure<'_>>::try_from(decoded).unwrap();
-        let outer = <(Str<'_>, Str<'_>, Structure<'_>)>::try_from(s).unwrap();
-        assert_eq!(outer.0, "a");
-        assert_eq!(outer.1, "b");
-
-        let inner = <(i32, i32)>::try_from(outer.2).unwrap();
-        assert_eq!(inner.0, 1);
-        assert_eq!(inner.1, 2);
-
-        #[derive(Serialize, Deserialize, Type, PartialEq, Debug)]
-        struct Foo {
-            val: u32,
-        }
-
-        let foo = Foo { val: 99 };
-        let v = as_value::Serialize(&foo);
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        let decoded: as_value::Deserialize<'_, Foo> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded.0, foo);
-
-        // Unit struct should be treated as a 0-sized tuple (the same as unit type)
-        #[derive(Serialize, Deserialize, Type, PartialEq, Debug)]
-        struct Unit;
-
-        assert_eq!(Unit::SIGNATURE, "");
-        let encoded = to_bytes(ctxt, &Unit).unwrap();
-        assert_eq!(encoded.len(), 0);
-        let _decoded: Unit = encoded.deserialize().unwrap().0;
-
-        // Structs w/o fields should be treated as a unit struct.
-        #[derive(Serialize, Deserialize, Type, PartialEq, Debug)]
-        struct NoFields {}
-
-        assert_eq!(NoFields::SIGNATURE, "y");
-        let encoded = to_bytes(ctxt, &NoFields {}).unwrap();
-        assert_eq!(encoded.len(), 1);
-        let _decoded: NoFields = encoded.deserialize().unwrap().0;
-
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(LE, 0);
-            let encoded = to_bytes(ctxt, &NoFields {}).unwrap();
-            assert_eq!(encoded.len(), 1);
-            let _decoded: NoFields = encoded.deserialize().unwrap().0;
-        }
-    }
-
-    #[test]
-    fn struct_ref() {
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &(&1u32, &2u32)).unwrap();
-        let decoded: [u32; 2] = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, [1u32, 2u32]);
-    }
-
-    #[test]
-    fn dict_value() {
-        use zvariant::{DeserializeDict, SerializeDict};
-
-        let mut map: HashMap<i64, &str> = HashMap::new();
-        map.insert(1, "123");
-        map.insert(2, "456");
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &map).unwrap();
-        assert_eq!(dbg!(encoded.len()), 40);
-        let decoded: HashMap<i64, &str> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded[&1], "123");
-        assert_eq!(decoded[&2], "456");
-
-        // GVariant format now
-        #[cfg(feature = "gvariant")]
-        {
-            let ctxt = Context::new_gvariant(NATIVE_ENDIAN, 0);
-            let gv_encoded = to_bytes(ctxt, &map).unwrap();
-            assert_eq!(gv_encoded.len(), 30);
-            let map: HashMap<i64, &str> = encoded.deserialize().unwrap().0;
-            assert_eq!(map[&1], "123");
-            assert_eq!(map[&2], "456");
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // As Value
-        let v: Value<'_> = Dict::from(map).into();
-        assert_eq!(v.value_signature(), "a{xs}");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 48);
-        // Convert it back
-        let dict: Dict<'_, '_> = v.try_into().unwrap();
-        let map: HashMap<i64, String> = dict.try_clone().unwrap().try_into().unwrap();
-        assert_eq!(map[&1], "123");
-        assert_eq!(map[&2], "456");
-        // Also decode it back
-        let v = encoded.deserialize().unwrap().0;
-        if let Value::Dict(dict) = v {
-            assert_eq!(dict.get::<i64, &str>(&1).unwrap().unwrap(), "123");
-            assert_eq!(dict.get::<i64, &str>(&2).unwrap().unwrap(), "456");
-        } else {
-            panic!();
-        }
-        // Convert it to a BTreeMap too.
-        let map: BTreeMap<i64, String> = dict.try_into().unwrap();
-        assert_eq!(map[&1], "123");
-        assert_eq!(map[&2], "456");
-        // Use iterator
-        let mut dict = Dict::from(map);
-        let expect = vec![
-            (Value::from(1i64), Value::from("123")),
-            (Value::from(2i64), Value::from("456")),
-        ];
-        let expect_iter = expect.iter().map(|(k, v)| (k, v)).collect::<Vec<_>>();
-        let actual = dict.iter().collect::<Vec<_>>();
-        assert_eq!(actual, expect_iter);
-        let actual = dict.iter().collect::<Vec<_>>();
-        assert_eq!(actual, expect_iter);
-        let actual = dict.iter().collect::<Vec<_>>();
-        assert_eq!(actual, expect_iter);
-        for (_, v) in dict.iter_mut() {
-            if let Value::Str(vv) = v {
-                *vv = Str::from(vv.to_string() + "-hello");
-            }
-        }
-        let actual = dict.into_iter().collect::<Vec<_>>();
-        let expect = vec![
-            (Value::from(1i64), Value::from("123-hello")),
-            (Value::from(2i64), Value::from("456-hello")),
-        ];
-        assert_eq!(actual, expect);
-
-        #[cfg(feature = "gvariant")]
-        {
-            // GVariant-format requires framing offsets for dict entries with variable-length keys
-            // so let's test that.
-            let mut map: HashMap<&str, &str> = HashMap::new();
-            map.insert("hi", "1234");
-            map.insert("world", "561");
-            let ctxt = Context::new_gvariant(NATIVE_ENDIAN, 0);
-            let gv_encoded = to_bytes(ctxt, &map).unwrap();
-            assert_eq!(gv_encoded.len(), 22);
-            let map: HashMap<&str, &str> = gv_encoded.deserialize().unwrap().0;
-            assert_eq!(map["hi"], "1234");
-            assert_eq!(map["world"], "561");
-
-            // Ensure SerializeValue produces the same result as Value
-            // Tests for https://github.com/dbus2/zbus/issues/868
-            let mut map = std::collections::HashMap::<&str, &str>::new();
-            map.insert("k", "v");
-            let gv_ser_value_encoded =
-                zvariant::to_bytes(ctxt, &as_value::Serialize(&map)).unwrap();
-            let gv_value_encoded = to_bytes(ctxt, &zvariant::Value::new(map)).unwrap();
-            assert_eq!(*gv_value_encoded, *gv_ser_value_encoded);
-
-            // Now the same but empty dict this time
-            let map: HashMap<&str, &str> = HashMap::new();
-            let gv_encoded = to_bytes(ctxt, &map).unwrap();
-            assert_eq!(gv_encoded.len(), 0);
-            let map: HashMap<&str, &str> = gv_encoded.deserialize().unwrap().0;
-            assert_eq!(map.len(), 0);
-        }
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // Now a hand-crafted Dict Value but with a Value as value
-        let mut dict = Dict::new(<&str>::SIGNATURE, Value::SIGNATURE);
-        dict.add("hello", Value::new("there")).unwrap();
-        dict.add("bye", Value::new("now")).unwrap();
-        let v: Value<'_> = dict.into();
-        assert_eq!(v.value_signature(), "a{sv}");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(dbg!(encoded.len()), 66);
-        let v: Value<'_> = encoded.deserialize().unwrap().0;
-        if let Value::Dict(dict) = v {
-            assert_eq!(
-                dict.get::<&str, Value<'_>>(&"hello").unwrap().unwrap(),
-                Value::new("there")
-            );
-            assert_eq!(
-                dict.get::<_, Value<'_>>(&"bye").unwrap().unwrap(),
-                Value::new("now")
-            );
-
-            // Try converting to a HashMap
-            let map = <HashMap<String, Value<'_>>>::try_from(dict.try_clone().unwrap()).unwrap();
-            assert_eq!(map["hello"], Value::new("there"));
-            assert_eq!(map["bye"], Value::new("now"));
-
-            // Try converting to a BTreeMap
-            let map = <BTreeMap<String, Value<'_>>>::try_from(dict).unwrap();
-            assert_eq!(map["hello"], Value::new("there"));
-            assert_eq!(map["bye"], Value::new("now"));
-        } else {
-            panic!();
-        }
-
-        #[derive(SerializeDict, DeserializeDict, Type, PartialEq, Debug, Default)]
-        #[zvariant(signature = "a{sv}")]
-        struct Test {
-            process_id: Option<u32>,
-            group_id: Option<u32>,
-            user: String,
-        }
-
-        let test = Test {
-            process_id: Some(42),
-            group_id: None,
-            user: "me".to_string(),
-        };
-
-        let encoded = to_bytes(ctxt, &test).unwrap();
-        assert_eq!(encoded.len(), 51);
-
-        let decoded: HashMap<&str, Value<'_>> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded["process_id"], Value::U32(42));
-        assert_eq!(decoded["user"], Value::new("me"));
-        assert!(!decoded.contains_key("group_id"));
-
-        let decoded: Test = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, test);
-
-        #[derive(SerializeDict, DeserializeDict, Type, PartialEq, Debug)]
-        #[zvariant(signature = "a{sv}")]
-        struct TestMissing {
-            process_id: Option<u32>,
-            group_id: Option<u32>,
-            user: String,
-            quota: u8,
-        }
-        let decoded: Result<(TestMissing, _)> = encoded.deserialize();
-        assert_eq!(
-            decoded.unwrap_err(),
-            Error::Message("missing field `quota`".to_string())
-        );
-
-        #[derive(SerializeDict, DeserializeDict, Type, PartialEq, Debug, Default)]
-        #[zvariant(signature = "a{sv}")]
-        struct TestSkipUnknown {
-            process_id: Option<u32>,
-            group_id: Option<u32>,
-        }
-        let _: TestSkipUnknown = encoded.deserialize().unwrap().0;
-
-        #[derive(SerializeDict, DeserializeDict, Type, PartialEq, Debug, Default)]
-        #[zvariant(signature = "a{sv}", deny_unknown_fields)]
-        struct TestDenyUnknown {
-            process_id: Option<u32>,
-            group_id: Option<u32>,
-        }
-        let decoded: Result<(TestDenyUnknown, _)> = encoded.deserialize();
-        assert_eq!(
-            decoded.unwrap_err(),
-            Error::Message("unknown field `user`, expected `process_id` or `group_id`".to_string())
-        );
-
-        #[derive(Serialize, Deserialize, Type, PartialEq, Debug, Default)]
-        #[serde(default)]
-        #[zvariant(signature = "a{sv}")]
-        struct TestParseUnknown<'s> {
-            #[serde(with = "optional", skip_serializing_if = "Option::is_none")]
-            process_id: Option<u32>,
-            #[serde(with = "optional", skip_serializing_if = "Option::is_none")]
-            group_id: Option<u32>,
-            #[serde(flatten, borrow)]
-            rest: HashMap<&'s str, Value<'s>>,
-        }
-        let decoded: TestParseUnknown<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded.rest.len(), 1);
-        assert_eq!(decoded.rest["user"], Value::new("me").try_into().unwrap());
-
-        #[cfg(feature = "gvariant")]
-        {
-            let test = Test {
-                process_id: Some(42),
-                group_id: None,
-                user: "me".to_string(),
-            };
-
-            let ctxt = Context::new_gvariant(NATIVE_ENDIAN, 0);
-            let encoded = to_bytes(ctxt, &test).unwrap();
-            let _: Test = encoded.deserialize().unwrap().0;
-            let decoded: Result<(TestMissing, _)> = encoded.deserialize();
-            assert_eq!(
-                decoded.unwrap_err(),
-                Error::Message("missing field `quota`".to_string())
-            );
-            let _: TestSkipUnknown = encoded.deserialize().unwrap().0;
-            let decoded: Result<(TestDenyUnknown, _)> = encoded.deserialize();
-            assert_eq!(
-                decoded.unwrap_err(),
-                Error::Message(
-                    "unknown field `user`, expected `process_id` or `group_id`".to_string()
-                )
-            );
-        }
-    }
-
-    #[test]
-    fn dict_compare() {
-        // the order in which a dict has been constructed must not play a role
-        // https://github.com/dbus2/zbus/issues/484
-        let mut dict1 = Dict::new(<&str>::SIGNATURE, Value::SIGNATURE);
-        dict1.add("first", Value::new("value")).unwrap();
-        dict1.add("second", Value::new("value")).unwrap();
-
-        let mut dict2 = Dict::new(<&str>::SIGNATURE, Value::SIGNATURE);
-        dict2.add("second", Value::new("value")).unwrap();
-        dict2.add("first", Value::new("value")).unwrap();
-
-        assert_eq!(dict1, dict2);
-    }
-
-    #[test]
-    fn value_value() {
-        let ctxt = Context::new_dbus(BE, 0);
-        let encoded = to_bytes(ctxt, &0xABBA_ABBA_ABBA_ABBA_u64).unwrap();
-        assert_eq!(encoded.len(), 8);
-        assert_eq!(LE.read_u64(&encoded), 0xBAAB_BAAB_BAAB_BAAB_u64);
-        let decoded: u64 = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, 0xABBA_ABBA_ABBA_ABBA);
-
-        // Lie about there being bytes before
-        let ctxt = Context::new_dbus(LE, 2);
-        let encoded = to_bytes(ctxt, &0xABBA_ABBA_ABBA_ABBA_u64).unwrap();
-        assert_eq!(encoded.len(), 14);
-        let decoded: u64 = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, 0xABBA_ABBA_ABBA_ABBA_u64);
-        let ctxt = Context::new_dbus(LE, 0);
-
-        // As Value
-        let v: Value<'_> = 0xFEFE_u64.into();
-        assert_eq!(v.value_signature(), "t");
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 16);
-        let v = encoded.deserialize().unwrap().0;
-        assert_eq!(v, Value::U64(0xFEFE));
-
-        // And now as Value in a Value
-        let v = Value::Value(Box::new(v));
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        assert_eq!(encoded.len(), 16);
-        let v = encoded.deserialize().unwrap().0;
-        if let Value::Value(v) = v {
-            assert_eq!(v.value_signature(), "t");
-            assert_eq!(*v, Value::U64(0xFEFE));
-        } else {
-            panic!();
-        }
-
-        // Ensure Value works with other Serializer & Deserializer
-        let v: Value<'_> = 0xFEFE_u64.into();
-        let encoded = serde_json::to_string(&v).unwrap();
-        let v = serde_json::from_str::<Value<'_>>(&encoded).unwrap();
-        assert_eq!(v, Value::U64(0xFEFE));
-
-        // Now a test case for https://github.com/dbus2/zbus/issues/549
-        #[derive(Deserialize, Serialize, Debug, PartialEq)]
-        struct Data {
-            inner: OwnedValue,
-        }
-
-        let value = zvariant::Value::new("variant-value");
-        let inner = zvariant::StructureBuilder::new()
-            .add_field("value1".to_string())
-            .add_field("value2")
-            .append_field(zvariant::Value::new(value)) // let's try to get a variant
-            .build()
-            .unwrap()
-            .try_into()
-            .unwrap();
-
-        let data = Data { inner };
-        let as_json = serde_json::to_value(&data).unwrap();
-        let expected_json = json!(
-            {
-                "inner": {
-                    "signature": "(ssv)",
-                    "value": [
-                        "value1",
-                        "value2",
-                        {
-                            "signature": "s",
-                            "value": "variant-value"
-                        }
-                    ]
-                }
-            }
-        );
-        assert_eq!(expected_json, as_json);
-        let data_again: Data = serde_json::from_str(&as_json.to_string()).unwrap();
-        assert_eq!(data, data_again);
-    }
-
-    #[test]
-    fn enums() {
-        use serde::{Deserialize, Serialize};
-
-        #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-        enum Unit {
-            Variant1,
-            Variant2,
-            Variant3,
-        }
-
-        let ctxts_n_expected_lens = [
-            // Unit variants are encoded as u32 and that has the same encoding in both formats.
-            [
-                (Context::new_dbus(BE, 0), 4usize),
-                (Context::new_dbus(BE, 1), 7),
-                (Context::new_dbus(BE, 2), 6),
-                (Context::new_dbus(BE, 3), 5),
-                (Context::new_dbus(BE, 4), 4),
-            ],
-            #[cfg(feature = "gvariant")]
-            [
-                (Context::new_gvariant(BE, 0), 4usize),
-                (Context::new_gvariant(BE, 1), 7),
-                (Context::new_gvariant(BE, 2), 6),
-                (Context::new_gvariant(BE, 3), 5),
-                (Context::new_gvariant(BE, 4), 4),
-            ],
-        ];
-        for ctxts_n_expected_len in ctxts_n_expected_lens {
-            for (ctxt, expected_len) in ctxts_n_expected_len {
-                let encoded = to_bytes_for_signature(ctxt, "u", &Unit::Variant2).unwrap();
-                assert_eq!(encoded.len(), expected_len);
-                let decoded: Unit = encoded.deserialize_for_signature("u").unwrap().0;
-                assert_eq!(decoded, Unit::Variant2);
-            }
-        }
-
-        #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-        enum NewType<'s> {
-            Variant1(&'s str),
-            Variant2(&'s str),
-            Variant3(&'s str),
-        }
-
-        let ctxts_n_expected_lens = [
-            [
-                (Context::new_dbus(BE, 0), 14usize),
-                (Context::new_dbus(BE, 1), 21),
-                (Context::new_dbus(BE, 2), 20),
-                (Context::new_dbus(BE, 3), 19),
-                (Context::new_dbus(BE, 4), 18),
-            ],
-            #[cfg(feature = "gvariant")]
-            [
-                (Context::new_gvariant(BE, 0), 10usize),
-                (Context::new_gvariant(BE, 1), 13),
-                (Context::new_gvariant(BE, 2), 12),
-                (Context::new_gvariant(BE, 3), 11),
-                (Context::new_gvariant(BE, 4), 10),
-            ],
-        ];
-        for ctxts_n_expected_len in ctxts_n_expected_lens {
-            for (ctxt, expected_len) in ctxts_n_expected_len {
-                let encoded =
-                    to_bytes_for_signature(ctxt, "(us)", &NewType::Variant2("hello")).unwrap();
-                assert_eq!(encoded.len(), expected_len);
-                let decoded: NewType<'_> = encoded.deserialize_for_signature("(us)").unwrap().0;
-                assert_eq!(decoded, NewType::Variant2("hello"));
-            }
-        }
-
-        #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-        enum Structs {
-            Tuple(u8, u32),
-            Struct { y: u8, t: u32 },
-        }
-
-        let ctxts_n_expected_lens = [
-            [
-                (Context::new_dbus(BE, 0), 16usize),
-                (Context::new_dbus(BE, 1), 23),
-                (Context::new_dbus(BE, 2), 22),
-                (Context::new_dbus(BE, 3), 21),
-                (Context::new_dbus(BE, 4), 20),
-            ],
-            #[cfg(feature = "gvariant")]
-            [
-                (Context::new_gvariant(BE, 0), 12usize),
-                (Context::new_gvariant(BE, 1), 15),
-                (Context::new_gvariant(BE, 2), 14),
-                (Context::new_gvariant(BE, 3), 13),
-                (Context::new_gvariant(BE, 4), 12),
-            ],
-        ];
-        // TODO: Provide convenience API to create complex signatures
-        let signature = "(u(yu))";
-        for ctxts_n_expected_len in ctxts_n_expected_lens {
-            for (ctxt, expected_len) in ctxts_n_expected_len {
-                let encoded =
-                    to_bytes_for_signature(ctxt, signature, &Structs::Tuple(42, 42)).unwrap();
-                assert_eq!(encoded.len(), expected_len);
-                let decoded: Structs = encoded.deserialize_for_signature(signature).unwrap().0;
-                assert_eq!(decoded, Structs::Tuple(42, 42));
-
-                let s = Structs::Struct { y: 42, t: 42 };
-                let encoded = to_bytes_for_signature(ctxt, signature, &s).unwrap();
-                assert_eq!(encoded.len(), expected_len);
-                let decoded: Structs = encoded.deserialize_for_signature(signature).unwrap().0;
-                assert_eq!(decoded, Structs::Struct { y: 42, t: 42 });
-            }
-        }
-    }
-
-    #[test]
-    fn derive() {
-        use serde::{Deserialize, Serialize};
-        use serde_repr::{Deserialize_repr, Serialize_repr};
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Struct<'s> {
-            field1: u16,
-            field2: i64,
-            field3: &'s str,
-        }
-
-        assert_eq!(Struct::SIGNATURE, "(qxs)");
-        let s = Struct {
-            field1: 0xFF_FF,
-            field2: 0xFF_FF_FF_FF_FF_FF,
-            field3: "hello",
-        };
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 26);
-        let decoded: Struct<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, s);
-
-        #[derive(Deserialize, Serialize, Type)]
-        struct UnitStruct;
-
-        assert_eq!(UnitStruct::SIGNATURE, <()>::SIGNATURE);
-        let encoded = to_bytes(ctxt, &UnitStruct).unwrap();
-        assert_eq!(encoded.len(), 0);
-        let _: UnitStruct = encoded.deserialize().unwrap().0;
-
-        #[repr(u8)]
-        #[derive(Deserialize_repr, Serialize_repr, Type, Value, OwnedValue, Debug, PartialEq)]
-        enum Enum {
-            Variant1,
-            Variant2,
-            Variant3,
-        }
-
-        assert_eq!(Enum::SIGNATURE, u8::SIGNATURE);
-        let encoded = to_bytes(ctxt, &Enum::Variant3).unwrap();
-        assert_eq!(encoded.len(), 1);
-        let decoded: Enum = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, Enum::Variant3);
-
-        assert_eq!(Value::from(Enum::Variant1), Value::U8(0));
-        assert_eq!(Enum::try_from(Value::U8(2)), Ok(Enum::Variant3));
-        assert_eq!(Enum::try_from(Value::U8(4)), Err(Error::IncorrectType));
-
-        #[repr(i64)]
-        #[derive(Deserialize_repr, Serialize_repr, Type, Value, OwnedValue, Debug, PartialEq)]
-        enum Enum2 {
-            Variant1,
-            Variant2,
-            Variant3,
-        }
-
-        assert_eq!(Enum2::SIGNATURE, i64::SIGNATURE);
-        let encoded = to_bytes(ctxt, &Enum2::Variant2).unwrap();
-        assert_eq!(encoded.len(), 8);
-        let decoded: Enum2 = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, Enum2::Variant2);
-
-        assert_eq!(Value::from(Enum2::Variant1), Value::I64(0));
-        assert_eq!(Enum2::try_from(Value::I64(2)), Ok(Enum2::Variant3));
-        assert_eq!(Enum2::try_from(Value::I64(4)), Err(Error::IncorrectType));
-
-        #[derive(Deserialize, Serialize, Type, Value, OwnedValue, Debug, PartialEq)]
-        enum NoReprEnum {
-            Variant1,
-            Variant2,
-            Variant3,
-        }
-
-        // issue#265: Panic on deserialization of a structure w/ a unit enum as its last field.
-        let encoded = to_bytes(ctxt, &(NoReprEnum::Variant2,)).unwrap();
-        let _: (NoReprEnum,) = encoded.deserialize().unwrap().0;
-
-        assert_eq!(NoReprEnum::SIGNATURE, u32::SIGNATURE);
-        let encoded = to_bytes(ctxt, &NoReprEnum::Variant2).unwrap();
-        assert_eq!(encoded.len(), 4);
-        let decoded: NoReprEnum = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, NoReprEnum::Variant2);
-
-        #[derive(Deserialize, Serialize, Type, Value, OwnedValue, Debug, PartialEq)]
-        #[zvariant(signature = "s", rename_all = "snake_case")]
-        enum StrEnum {
-            VariantOne,
-            Variant2,
-            Variant3,
-        }
-
-        assert_eq!(StrEnum::SIGNATURE, <&str>::SIGNATURE);
-        let encoded = to_bytes(ctxt, &StrEnum::Variant2).unwrap();
-        assert_eq!(encoded.len(), 13);
-        let decoded: StrEnum = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, StrEnum::Variant2);
-
-        assert_eq!(
-            StrEnum::try_from(Value::Str("variant_one".into())),
-            Ok(StrEnum::VariantOne)
-        );
-        assert_eq!(
-            StrEnum::try_from(Value::Str("variant2".into())),
-            Ok(StrEnum::Variant2)
-        );
-        assert_eq!(
-            StrEnum::try_from(Value::Str("variant4".into())),
-            Err(Error::IncorrectType)
-        );
-        assert_eq!(StrEnum::try_from(Value::U32(0)), Err(Error::IncorrectType));
-
-        #[derive(Deserialize, Serialize, Type)]
-        enum NewType {
-            Variant1(f64),
-            Variant2(f64),
-        }
-        assert_eq!(NewType::SIGNATURE, "(ud)");
-
-        #[derive(Deserialize, Serialize, Type)]
-        enum StructFields {
-            Variant1(u16, i64, &'static str),
-            Variant2 {
-                field1: u16,
-                field2: i64,
-                field3: &'static str,
-            },
-        }
-        assert_eq!(StructFields::SIGNATURE, "(u(qxs))");
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct AStruct<'s> {
-            field1: u16,
-            field2: &'s [u8],
-            field3: &'s [u8],
-            field4: i64,
-        }
-        assert_eq!(AStruct::SIGNATURE, "(qayayx)");
-        let s = AStruct {
-            field1: 0xFF_FF,
-            field2: &[77u8; 8],
-            field3: &[77u8; 8],
-            field4: 0xFF_FF_FF_FF_FF_FF,
-        };
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 40);
-        let decoded: AStruct<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, s);
-    }
-
-    #[test]
-    fn serialized_size() {
-        let ctxt = Context::new_dbus(LE, 0);
-        let l = crate::serialized_size(ctxt, &()).unwrap();
-        assert_eq!(*l, 0);
-
-        #[cfg(unix)]
-        {
-            let stdout = std::io::stdout();
-            let l = crate::serialized_size(ctxt, &Fd::from(&stdout)).unwrap();
-            assert_eq!(*l, 4);
-            assert_eq!(l.num_fds(), 1);
-        }
-
-        let l = crate::serialized_size(ctxt, &('a', "abc", &(1_u32, 2))).unwrap();
-        assert_eq!(*l, 24);
-
-        let v = vec![1, 2];
-        let l = crate::serialized_size(ctxt, &('a', "abc", &v)).unwrap();
-        assert_eq!(*l, 28);
-    }
-
-    #[test]
-    #[cfg(feature = "serde_bytes")]
-    fn serde_bytes() {
-        use serde::{Deserialize, Serialize};
-        use serde_bytes::*;
-
-        let ctxt = Context::new_dbus(LE, 0);
-        let ay = Bytes::new(&[77u8; 1_000_000]);
-        let encoded = to_bytes(ctxt, &ay).unwrap();
-        assert_eq!(encoded.len(), 1_000_004);
-        let decoded: ByteBuf = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded.len(), 1_000_000);
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Struct<'s> {
-            field1: u16,
-            #[serde(with = "serde_bytes")]
-            field2: &'s [u8],
-            field3: i64,
-        }
-        assert_eq!(Struct::SIGNATURE, "(qayx)");
-        let s = Struct {
-            field1: 0xFF_FF,
-            field2: &[77u8; 512],
-            field3: 0xFF_FF_FF_FF_FF_FF,
-        };
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 528);
-        let decoded: Struct<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, s);
-    }
-
-    #[test]
-    #[cfg(all(feature = "serde_bytes", feature = "gvariant"))]
-    fn serde_bytes_gvariant() {
-        use serde::{Deserialize, Serialize};
-        use serde_bytes::*;
-
-        let ctxt = Context::new_gvariant(LE, 0);
-        let ay = Bytes::new(&[77u8; 1_000_000]);
-        let encoded = to_bytes(ctxt, &ay).unwrap();
-        assert_eq!(encoded.len(), 1_000_000);
-        let decoded: ByteBuf = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded.len(), 1_000_000);
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Struct<'s> {
-            field1: u16,
-            #[serde(with = "serde_bytes")]
-            field2: &'s [u8],
-            field3: i64,
-        }
-        assert_eq!(Struct::SIGNATURE, "(qayx)");
-        let s = Struct {
-            field1: 0xFF_FF,
-            field2: &[77u8; 512],
-            field3: 0xFF_FF_FF_FF_FF_FF,
-        };
-        let encoded = to_bytes(ctxt, &s).unwrap();
-        assert_eq!(encoded.len(), 530);
-        let decoded: Struct<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, s);
-    }
-
-    #[test]
-    #[cfg(any(feature = "gvariant", feature = "option-as-array"))]
-    fn option_value() {
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        let ctxt = Context::new_gvariant(NATIVE_ENDIAN, 0);
-        #[cfg(feature = "option-as-array")]
-        let ctxt = Context::new_dbus(NATIVE_ENDIAN, 0);
-
-        // First a Some fixed-sized value
-        let mn = Some(16i16);
-        let encoded = to_bytes(ctxt, &mn).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 2);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 6);
-        let decoded: Option<i16> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, mn);
-
-        // As Value
-        let v: Value<'_> = mn.into();
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 5);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 10);
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        match decoded {
-            #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-            Value::Maybe(maybe) => assert_eq!(maybe.get().unwrap(), mn),
-            #[cfg(feature = "option-as-array")]
-            Value::Array(array) => {
-                assert_eq!(i16::try_from(array[0].try_clone().unwrap()).unwrap(), 16i16)
-            }
-            _ => panic!("unexpected value {decoded:?}"),
-        }
-
-        // Now a None of the same type
-        let mn: Option<i16> = None;
-        let encoded = to_bytes(ctxt, &mn).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 0);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 4);
-        let decoded: Option<i16> = encoded.deserialize().unwrap().0;
-        assert!(decoded.is_none());
-
-        // Next a Some variable-sized value
-        let ms = Some("hello world");
-        let encoded = to_bytes(ctxt, &ms).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 13);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 20);
-        let decoded: Option<&str> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, ms);
-
-        // As Value
-        let v: Value<'_> = ms.into();
-        #[cfg(feature = "option-as-array")]
-        match &v {
-            Value::Array(array) => {
-                assert_eq!(
-                    String::try_from(array[0].try_clone().unwrap()).unwrap(),
-                    ms.unwrap()
-                )
-            }
-            _ => panic!("unexpected value {v:?}"),
-        }
-
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 16);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 24);
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        match decoded {
-            #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-            Value::Maybe(maybe) => {
-                assert_eq!(maybe.get::<String>().unwrap().as_deref(), ms);
-            }
-            #[cfg(feature = "option-as-array")]
-            Value::Array(array) => {
-                assert_eq!(
-                    String::try_from(array[0].try_clone().unwrap()).unwrap(),
-                    ms.unwrap()
-                )
-            }
-            _ => panic!("unexpected value {decoded:?}"),
-        }
-
-        // Now a None of the same type
-        let ms: Option<&str> = None;
-        let encoded = to_bytes(ctxt, &ms).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 0);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 4);
-        let decoded: Option<&str> = encoded.deserialize().unwrap().0;
-        assert!(decoded.is_none());
-
-        // In a seq type
-        let ams = vec![
-            Some(String::from("hello world")),
-            Some(String::from("bye world")),
-        ];
-        let encoded = to_bytes(ctxt, &ams).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 26);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 42);
-        let decoded: Vec<Option<String>> = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, ams);
-
-        // As Value
-        let v: Value<'_> = ams.clone().into();
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 30);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 50);
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(v, decoded);
-
-        // In a struct
-        let structure: (Option<String>, u64, Option<String>) =
-            (Some(String::from("hello world")), 42u64, None);
-        let encoded = to_bytes(ctxt, &structure).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 25);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 36);
-        let decoded: (Option<String>, u64, Option<String>) = encoded.deserialize().unwrap().0;
-        assert_eq!(decoded, structure);
-
-        // As Value
-        let v: Value<'_> = structure.clone().into();
-        let encoded = to_bytes(ctxt, &v).unwrap();
-        #[cfg(all(feature = "gvariant", not(feature = "option-as-array")))]
-        assert_eq!(encoded.len(), 33);
-        #[cfg(feature = "option-as-array")]
-        assert_eq!(encoded.len(), 52);
-        let decoded: Value<'_> = encoded.deserialize().unwrap().0;
-        assert_eq!(v, decoded);
-    }
-
-    #[test]
-    fn struct_with_hashmap() {
-        use serde::{Deserialize, Serialize};
-
-        let mut hmap = HashMap::new();
-        hmap.insert("key".into(), "value".into());
-
-        #[derive(Type, Deserialize, Serialize, PartialEq, Debug)]
-        struct Foo {
-            hmap: HashMap<String, String>,
-        }
-
-        let foo = Foo { hmap };
-        assert_eq!(Foo::SIGNATURE, "(a{ss})");
-
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &(&foo, 1)).unwrap();
-        let f: Foo = encoded.deserialize().unwrap().0;
-        assert_eq!(f, foo);
-    }
-
-    #[test]
-    fn issue_59() {
-        // Ensure we don't panic on deserializing tuple of smaller than expected length.
-        let ctxt = Context::new_dbus(LE, 0);
-        let encoded = to_bytes(ctxt, &("hello",)).unwrap();
-        let result: Result<((&str, &str), _)> = encoded.deserialize();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    #[cfg(feature = "gvariant")]
-    fn issue_99() {
-        use crate::to_bytes_for_signature;
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct ZVStruct<'s>(#[serde(borrow)] HashMap<&'s str, Value<'s>>);
-
-        let mut dict = HashMap::new();
-        dict.insert("hi", Value::from("hello"));
-        dict.insert("bye", Value::from("then"));
-
-        let element = ZVStruct(dict);
-
-        let ctxt = Context::new_gvariant(LE, 0);
-        let signature = ZVStruct::SIGNATURE;
-
-        let encoded = to_bytes_for_signature(ctxt, signature, &element).unwrap();
-        let _: ZVStruct<'_> = encoded.deserialize_for_signature(signature).unwrap().0;
-    }
-
-    #[test]
-    fn issue_1145() {
-        // Ensure f32::NAN can be encoded and decoded.
-        let ctxt = Context::new_dbus(LE, 0);
-        {
-            let encoded = to_bytes(ctxt, &f32::NAN).unwrap();
-            let result: f32 = encoded.deserialize().unwrap().0;
-            assert!(result.is_nan());
-        }
-        // Ensure f32::INFINITY can be encoded and decoded.
-        {
-            let encoded = to_bytes(ctxt, &f32::INFINITY).unwrap();
-            let result: f32 = encoded.deserialize().unwrap().0;
-            assert!(result.is_infinite());
-        }
-        {
-            let encoded = to_bytes(ctxt, &f32::NEG_INFINITY).unwrap();
-            let result: f32 = encoded.deserialize().unwrap().0;
-            assert!(result.is_infinite());
-        }
-    }
-
-    #[cfg(feature = "ostree-tests")]
-    #[test]
-    fn ostree_de() {
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Summary<'a>(Vec<Repo<'a>>, #[serde(borrow)] HashMap<&'a str, Value<'a>>);
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Repo<'a>(&'a str, #[serde(borrow)] Metadata<'a>);
-
-        #[derive(Deserialize, Serialize, Type, PartialEq, Debug)]
-        struct Metadata<'a>(u64, Vec<u8>, #[serde(borrow)] HashMap<&'a str, Value<'a>>);
-
-        let encoded = std::fs::read("../test-data/flatpak-summary.dump").unwrap();
-        let ctxt = Context::new_gvariant(LE, 0);
-        let encoded = crate::serialized::Data::new(encoded, ctxt);
-        let _: Summary<'_> = encoded.deserialize().unwrap().0;
-        // If we're able to deserialize all the data successfully, don't bother checking the summary
-        // data.
-    }
-
-    #[test]
-    fn recursion_limits() {
-        let ctxt = Context::new_dbus(LE, 0);
-        // Total container depth exceeds limit (64)
-        let mut value = Value::from(0u8);
-        for _ in 0..64 {
-            value = Value::Value(Box::new(value));
-        }
-        assert!(matches!(
-            to_bytes(ctxt, &value),
-            Err(Error::MaxDepthExceeded(MaxDepthExceeded::Container))
-        ));
-
-        // Array depth exceeds limit (32)
-        let vec = vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![
-            vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![
-                vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![vec![
-                    0u8,
-                ]]]]]]]]]]],
-            ]]]]]]]]]]],
-        ]]]]]]]]]]];
-        assert!(matches!(
-            to_bytes(ctxt, &vec),
-            Err(Error::MaxDepthExceeded(MaxDepthExceeded::Array))
-        ));
-
-        // Struct depth exceeds limit (32)
-        let tuple = ((((((((((((((((((((((
-            (((((((((((0u8,),),),),),),),),),),),
-        ),),),),),),),),),),),),),),),),),),),),),);
-        assert!(matches!(
-            to_bytes(ctxt, &tuple),
-            Err(Error::MaxDepthExceeded(MaxDepthExceeded::Structure))
-        ));
-
-        // total depth exceeds limit (64) with struct, array and variant.
-        let mut value = Value::from(0u8);
-        for _ in 0..32 {
-            value = Value::Value(Box::new(value));
-        }
-        let tuple_array =
-            (
-                ((((((((((((((((vec![vec![vec![vec![vec![vec![vec![vec![
-                    vec![vec![vec![vec![vec![vec![vec![vec![value]]]]]]]],
-                ]]]]]]]],),),),),),),),),),),),),),),),),
-            );
-        assert!(matches!(
-            to_bytes(ctxt, &tuple_array),
-            Err(Error::MaxDepthExceeded(MaxDepthExceeded::Container))
-        ));
-
-        // TODO:
-        //
-        // * Test deserializers.
-        // * Test gvariant format.
-    }
-}
