+++ zvariant/tests/issue/mod.rs
@@ -1,0 +1,4 @@
+mod issue_1145;
+mod issue_59;
+#[cfg(feature = "gvariant")]
+mod issue_99;
