+++ zvariant/tests/number/i64_value.rs
@@ -1,0 +1,9 @@
+use zvariant::{BE, LE};
+
+#[test]
+fn i64_value() {
+    let encoded = basic_type_test!(BE, DBus, -0xABBA_ABBA_ABBA_AB0_i64, 8, i64, 8, I64, 16);
+    assert_eq!(LE.read_i64(&encoded), 0x5055_4455_4455_44F5_i64);
+    #[cfg(feature = "gvariant")]
+    basic_type_test!(BE, GVariant, -0xABBA_ABBA_ABBA_AB0_i64, 8, i64, 8, I64, 10);
+}
