+++ zvariant/tests/number/mod.rs
@@ -1,0 +1,13 @@
+#[macro_use]
+pub(self) mod common {
+    include!("../common.rs");
+}
+
+mod f64_value;
+mod i16_value;
+mod i32_value;
+mod i64_value;
+mod i8_value;
+mod u16_value;
+mod u32_value;
+mod u8_value;
